requirements.txt
streamlit
torch
opencv-python
numpy
onnxruntime
Pillow
ultralytics
